C Maximum of all Subarray of dim K
Complete (ready for testing) version 1 of "Maximum of all Subarrays of dimension k"
This is my first version in C++, using the deque from the STL; there is a part which is not efficient: guess which one? The part in which we erase the element going out from the subarray: in this version we have to look for it. 
This is because the pure deque structure is not optimal, we would need a doubly pointed one:
that is, it would suffice to save the pointer or reference of the element in a mod K array  (lenght k) to directly access them, so this would add an array of dimension k in the space needed; moreover you would need to save in the deque the pointer to the element in such array instead than the simple index which I am recording in the current version.
I save the temporary results in a sstream to cout them later, after all the cin s.